Move LIFDetectorsMHL and PureLIFDetectorsMHL out of lutorch into experiment folder
Retired variants kept as experiment-local code (LIFDetectorsMHL = the P_ij ordered-pair
0.64 champion; PureLIFDetectorsMHL = the M=2 bit-address variant subsumed by
ProductBucketLIFMHL). git mv preserves history; imports in distill_walker2d.py /
distill_walker2d_ttfs.py and the two moved tests updated to the local same-dir path;
tests green (16 passed).
