+++ experiments/lif_detectors_mhl/pure_lif_detectors_mhl.py
